Let the device tell us whether it's using the T or CS package.
This actually started with debugging the fact that my TSL2561 returned 0x0A in
its ID register when it was off but 0x00 when it was on.  I subsequently
realized two things: the bits are reversed from what the datasheet says, and
I can use the ID register to distinguish between the T and CS packages, so that
doesn't need to be an initialization parameter.